requirements.txt
dj-database-url==0.4.1
gunicorn==19.6.0
whitenoise==2.0.6
dj-static==0.0.6
autopep8==1.4.3
django-tinymce==2.3.0
html5lib==0.9999999
pep8==1.7.0
static3==0.7.0
openpyxl==2.4.5
xhtml2pdf==0.0.6
django-celery==3.2.1
celery==3.1.25
boto == 2.47.0
boto3 == 1.4.4
django-storages == 1.6.3
django-debug-toolbar==1.11
django-redis==4.10.0
djangorestframework==3.9.4
redis==3.3.11
django-related-admin==0.7.0
requests==2.31.0

asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
pillow==9.5.0
psycopg2==2.9.9
pyodbc==5.0.0b3
pypyodbc==1.3.6
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2024.1